Add stars for endorsed packages in the package chooser
